+++ backend/sensors/sensors.go
@@ -1,0 +1,1 @@
+package sensors
